Update patch set 2
Patch Set 2: (1 inline comment)

My first impression, when I started working with this, was that the enforcement status would be in the "security" tab of the settings app, not in "about phone".  Other people I talked with had the same reaction.  I'm still not sure which makes more sense.

Patch-set: 2