Update patch set 2
Patch Set 2:

It sounds to me that the deeper issue is either that CTS does something it shouldn't do and/or that the device policy lets it do so.

I'm for defense in depth, but I'm also for fixing things twice: let's make sure that the same condition won't cause a crash again, but let's also make sure that the condition doesn't happen again.

Patch-set: 2
Label: Code-Review=0
Label: Verified=0