Update patch set 5
Patch Set 5:

More detail on translated strings:

An example of the use of a translated string is at src/com/android/settings/deviceinfo/Status.java:

sUnknown = mRes.getString(R.string.device_info_default);

The getString() call retrieves the string value for the current language.  For the Ukranian example, this resolves to:

<string name="device_info_default" msgid="7847265875578739287">"Невідомо"</string>

Google translate says that "SELinux status" is "SELinux статус" in Ukranian.  So assuming that's correct, I guess you would add

<string name="selinux_status" msgid="YYYY">"SELinux статус"</string>

There is a hint on how to generate the msgid at http://stackoverflow.com/questions/4743058/whats-the-meaning-of-attribute-msgid-in-strings-xml :

"one trick is to take strings.xml from non-default language, copy it to raw/assets dir and make a build with it. Then read it programatically from raw/assets and use xml parser to extract the value you need"

Patch-set: 5