Update patch set 1
Patch Set 1:

Hmm.  I understand that the RSSI timing and scan timing will be different.  I would expect though that when the status-bar level changes the connected AP in the list would also change level.

What I saw yesterday in testing at a different location was that the sig level of the connected AP never changed in the list, even when the status-bar level changed.

What I see today in my new testing location is that it *is* changing correctly.  Sometimes it's off by a little bit - I wonder if the status bar uses a different method of calculating level?

I will have to investigate and get back to you.

Thank you very much for the change though - it is a big improvement.

Patch-set: 1