Update patch set 4
Patch Set 4:

Hi Björn,

I can surely look into this and pass your feedback on to our reviewer (which is always the best thing to do to check validity of a translation) but the translation I find in our translation console for "Time on" is "Tid aktiverad".
there are no instances of "Tid funktionen varit på".
And the same goes for "Not in range": the translation in TC is "Utom räckhåll".

It looks like there are some concerns about Swedish translation quality. Can you please loop me in those bugs so we can check on our side too? 

Thanks
Angela

Patch-set: 4