Update patch set 4
Patch Set 4:

Yes, there are indeed some concerns about the Swedish translation quality in general in Android. I'll loop you on the ones I find but I don't actively look for them and I set my phone to English normally, partly due to the inconsistent quality of the translations.

I think the general problem is that Android introduces many one-word strings into the translations and that these are just initially translated using Google Translate or equivilent automatic translation. Unless the person manually verifying the results is familiar with the context in which the specific translated word is used it becomes very hard to find the issues for those short translations.

Patch-set: 4