Update patch set 3
Patch Set 3: Code-Review+1

Added (back) the 16MB setting because in the logd+init review we declared 3% individual memory impact as our maximum, and would allow 16MB on a svelte device as a result.

NB: we do not pre-allocate the buffer, this is an impact only felt over a long period of time, and only if the logs in all 5 categories filled up to the brim.

Patch-set: 3
Label: Code-Review=+1