Update patch set 3
Patch Set 3:

> (1 comment)

That makes sense. However we need to consider translations. The problem is that app_list_memory has character limit of 30 and some translations that we would have to use if we switch would be longer than that since memory_usage_apps has a bit longer limit, 40 characters. (I guess potential problem exists in the current proposal as well.) Suggestions? Or is this not a problem?

Patch-set: 3
Reviewer: Zoran Jovanovic <1010077@85c56323-6fa9-3386-8a01-6480fb634889>