Update patch set 3
Patch Set 3:

> we suggested A2DP to be high priority is because there are also
 > many in-ear mono handsfree devices that support HFP but actually
 > can not provide stereo sound to user. this is also the reason SOMC
 > uses Mono type icon for ic_bt_headset_hfp.png.

Okay, but in AOSP we have a ic_bt_headset_hfp that looks like a stereo headset.  I want to avoid showing an icon that implies you are not using the microphone on the device.

A more acceptable solution would add a third icon that is used when a device supports both a2dp and hfp, which can indicate clearly that a stereo output is possible with a microphone.

Patch-set: 3
Reviewer: Marie Janssen <1073753@85c56323-6fa9-3386-8a01-6480fb634889>