Update patch set 3
Patch Set 3:

I'm missing a little bit of context in this CL. I understand that you could leak by having the activity as the first object to create the index. However, this doesn't look like safe singleton creation (multithreading will be an issue w/o locking).

Can you mark the remaining constructor @visiblefortesting if its only used for testing? If its not, then you need to refactor to make sure that you're consistently accessing the index via the singleton or add documentation as to why you might have a singleton index as well as an instance.

This also brings up the idea that the context is swappable, but you're alway swapping in the application context of the passed object, so that doesn't seem to have a lot of value there.

Patch-set: 3